# Peken-Banyumasan-Gate-Masuk-Keluar/openapi-gate.yaml
openapi: 3.1.0
info:
  title: Peken Banyumasan — Gate (Admin & Petugas) API
  version: 2.4.1
  description: |
    API consumed by the Gate admin/petugas panel
    (`Peken-Banyumasan-Gate-Masuk-Keluar/frontend/`).

    **Auth (v2.0 — Supabase):**
    - Login/logout/me/changePassword are **Supabase-direct** — frontend calls
      `supabase.auth.signInWithPassword()` etc. without touching this backend.
    - Gate has **no self-register** — admin and petugas accounts are created by superadmin.
    - `PUT /api/auth/profile` (custom fields only) is **HYBRID**.
    - All other endpoints require a valid **Supabase JWT** in `Authorization: Bearer <token>`.
      Backend verifies using `SUPABASE_JWT_SECRET` and enforces `app_metadata.role IN ('admin', 'petugas')`.
    - Endpoints exclusive to `role: admin` are noted in their summary. Petugas is limited to
      dashboard, visitor scan, and basic event/zone read.

    **Vocabulary (CANONICAL — non-negotiable):**
    - `kolaborator` (NEVER member, kreator)
    - `artisan` (NEVER tenant, umkm)
    - `aktivitas` for admin-moderation story feed (NEVER admin/stories)
    - Path segment `/kolaborator` singular (NEVER `/kolaborators`)
    - Path segment `/artisan` singular (NEVER `/artisans`)

    ---

    **Response Envelope (ALL endpoints except binary file downloads):**
    ```
    Success: { "status": "success", "message": "<optional>", "data": <payload> }
...
    ```
...

    **Authentication:**
...
    does NOT need to implement those endpoints.

...

...

...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /api/reports/artisan:
    get:
      operationId: getArtisanReport
      tags: [reports]
      summary: Get artisan revenue report (admin only)
      parameters:
        - name: event_id
          in: query
          schema:
            type: string
          description: Filter by event (optional — all events if omitted)
      responses:
        '200':
          description: Artisan report rows
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/ArtisanReportRow'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /api/events/{id}/artisan:
    get:
      operationId: listEventArtisan
      tags: [events-artisan]
      summary: List artisans assigned to event (admin only)
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: EventArtisan list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/EventArtisan'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      operationId: assignArtisanToEvent
      tags: [events-artisan]
      summary: Admin-assign artisan to event (admin only)
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [artisan_id]
              properties:
                artisan_id:
                  type: string
                stand_id:
                  type: string
                  nullable: true
                  description: Concrete zone stand (validated against the event's zones)
                posisi_event:
                  type: string
                  nullable: true
                  description: Free-text position label (e.g. "Zona A - Stand 5"); falls back to stand_id when omitted
                status_request:
                  type: string
                  enum: [approved]
                  default: approved
            example:
              artisan_id: art-0001
              posisi_event: Zona A - Stand 5
              status_request: approved
      responses:
        '200':
          description: Assigned
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/EventArtisan'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/events/{id}/artisan/{aid}:
    patch:
      operationId: updateEventArtisan
      tags: [events-artisan]
      summary: Update artisan assignment in event (admin only)
      parameters:
        - $ref: '#/components/parameters/IdPath'
        - name: aid
          in: path
          required: true
          schema:
            type: string
          description: EventArtisan junction ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                stand_id:
                  type: string
                  nullable: true
                posisi_event:
                  type: string
                  nullable: true
                status_request:
                  type: string
                  enum: [pending, pending_change, approved, rejected]
            example:
              stand_id: B-2
              status_request: approved
      responses:
        '200':
          description: Updated assignment
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/EventArtisan'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      operationId: removeEventArtisan
      tags: [events-artisan]
      summary: Remove artisan from event (admin only)
      parameters:
        - $ref: '#/components/parameters/IdPath'
        - name: aid
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Removed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope'
              example:
                status: success
                message: Artisan dihapus dari event
                data: null
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/artisan:
    get:
      operationId: listArtisan
      tags: [artisan-management]
      summary: List all artisans (admin only)
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [aktif, pending, suspended, rejected]
        - name: q
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Artisan list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Artisan'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/artisan/{id}:
    get:
      operationId: getArtisan
      tags: [artisan-management]
      summary: Get artisan detail (admin only)
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Artisan record
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Artisan'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      operationId: patchArtisan
      tags: [artisan-management]
      summary: Admin update artisan record (admin only — can edit komisi_persen)
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                komisi_persen:
                  type: number
                  minimum: 0
                  maximum: 100
                nama_usaha:
                  type: string
                pemilik:
                  type: string
                email:
                  type: string
                  format: email
                no_hp:
                  type: string
                kota:
                  type: string
                deskripsi:
                  type: string
                kategori_usaha:
                  type: array
                  items:
                    type: string
                foto_url:
                  type: string
                  nullable: true
                cover_url:
                  type: string
                  nullable: true
                internal_notes:
                  type: string
            example:
              komisi_persen: 12
              internal_notes: Awaiting QRIS verification.
      responses:
        '200':
          description: Updated artisan
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Artisan'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/artisan/{id}/status:
    patch:
      operationId: patchArtisanStatus
      tags: [artisan-management]
      summary: Update artisan status (admin only)
      description: |
        **Dual update required:** `app_metadata.status` in Supabase Auth (Admin SDK)
        AND `status` in `artisans` table, atomically.
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [aktif, suspended, rejected]
            example:
              status: aktif
      responses:
        '200':
          description: Status updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/artisan/{id}/events:
    get:
      operationId: listArtisanEvents
      tags: [artisan-management]
      summary: List events an artisan has joined (admin, read-only)
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Events list (event rows the artisan has joined)
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/ArtisanEventEntry'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/artisan/{id}/kas:
    get:
      operationId: getArtisanKas
      tags: [artisan-management]
      summary: Get artisan cashflow (admin, read-only)
      parameters:
        - $ref: '#/components/parameters/IdPath'
        - name: from
          in: query
          schema:
            type: string
            format: date
        - name: to
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Kas list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Kas'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/artisan/{id}/riwayat:
    get:
      operationId: getArtisanRiwayat
      tags: [artisan-management]
      summary: Get artisan transaction history (admin, read-only)
      parameters:
        - $ref: '#/components/parameters/IdPath'
        - name: from
          in: query
          schema:
            type: string
            format: date
        - name: to
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Riwayat list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Riwayat'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/artisan/{id}/promo:
    get:
      operationId: getArtisanPromo
      tags: [artisan-management]
      summary: Get artisan promos (admin, read-only)
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Promo list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Promo'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/artisan/{id}/stok:
    get:
      operationId: getArtisanStok
      tags: [artisan-management]
      summary: Get artisan inventory (admin, read-only)
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Stok list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Stok'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/artisan/{id}/qris:
    get:
      operationId: getArtisanQris
      tags: [artisan-management]
      summary: Get artisan QRIS URL (admin, read-only)
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: QRIS info
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          qris_url:
                            type: string
                            nullable: true
                          updated_at:
                            type: string
                            format: date-time
                            nullable: true
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  # ── Aktivitas (Admin Story Moderation) ────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Supabase JWT (RS256). Issued by `supabase.auth.signInWithPassword()` on the frontend.
        Backend verifies using `SUPABASE_JWT_SECRET`.
        Role must be 'admin' OR 'petugas' in `app_metadata.role`.
        Endpoints marked "admin only" additionally reject role 'petugas' with 403.
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Resource UUID or prefixed short ID
      example: evt-0001
  responses:
    Unauthorized:
      description: Missing, invalid, or expired Supabase JWT
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            status: error
            message: Sesi Anda telah berakhir
            data: null
    Forbidden:
      description: Authenticated but role insufficient (e.g. petugas accessing admin-only endpoint)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            status: error
            message: Akses ditolak
            data: null
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            status: error
            message: Sumber daya tidak ditemukan
            data: null
    Conflict:
      description: Conflict (e.g. stand already occupied)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            status: error
            message: Stand sudah ditempati artisan lain
            data: null
...
  schemas:
    Envelope:
      type: object
      required: [status, data]
      properties:
        status:
          type: string
          enum: [success]
        message:
          type: string
          nullable: true
        data:
          description: Actual payload — shape varies per endpoint
    ErrorEnvelope:
      type: object
      required: [status, message, data]
      properties:
        status:
          type: string
          enum: [error]
        message:
          type: string
        data:
          type: object
          nullable: true
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
...
    ArtisanReportRow:
      type: object
      required: [id, nama_usaha, kategori, omset, komisi_persen, transaksi]
      properties:
        id:
          type: string
          example: art-0001
        nama_usaha:
          type: string
          example: Batik Budi
        kategori:
          type: string
          description: Display string (joined kategori_usaha, e.g. 'Kriya, Fashion')
          example: F&B / Kuliner, Kriya
        omset:
          type: number
          description: Total revenue (rupiah)
          example: 5400000
        komisi_persen:
          type: number
          example: 10
        transaksi:
          type: integer
          description: Total transaction count
          example: 36
        event_count:
          type: integer
          description: Total events participated
          example: 3
        stand_terakhir:
          type: string
          nullable: true
          description: Last assigned stand ID
          example: A-3
...
    EventArtisan:
      type: object
      description: |
        Approved artisan participation row.
        Fields nama_usaha and kategori_usaha are JOINed from artisans table
        (NOT stored in event_artisans junction table).
      required: [id, nama_usaha, kategori_usaha, status_request, assigned_by]
      properties:
        id:
          type: string
          example: evart-0001
        event_id:
          type: string
          example: evt-0001
        artisan_id:
          type: string
          example: art-0001
        nama_usaha:
          type: string
          example: Batik Budi
        kategori_usaha:
          type: array
          items:
            type: string
          description: UMKM 9 options (mirrored from Artisan)
          example: ['F&B / Kuliner', 'Kriya']
        stand_id:
          type: string
          nullable: true
          description: Assigned stand (e.g. 'A-3')
          example: A-3
        posisi_event:
          type: string
          nullable: true
          description: Position/role within the event
          example: Stand A-3
        status_request:
          type: string
          enum: [pending, pending_change, approved, rejected]
          example: approved
        assigned_by:
          type: string
          enum: [admin, self]
          example: admin
    ArtisanEventEntry:
      type: object
      description: |
        Artisan-side view: one event an artisan has joined
        (GET /api/artisan/{id}/events). nama + date/time/location/status are
        JOINed from the events table; stand_id/posisi_event/status_request/
        assigned_by come from the event_artisans junction. id is the
        junction-row id.
      required: [id, event_id, artisan_id, nama, tanggal, tanggal_selesai, jam_mulai, jam_selesai, lokasi, status, status_request, assigned_by, created_at]
      properties:
        id:
          type: string
          description: Junction-row id (event_artisans.id)
          example: evart-0001
        event_id:
          type: string
          example: evt-0001
        artisan_id:
          type: string
          example: art-0001
        nama:
          type: string
          description: Event name (JOINed from events)
          example: Peken Banyumas #7
        tanggal:
          type: string
          format: date
          description: Event start date (WIB local YYYY-MM-DD)
          example: '2026-05-10'
        tanggal_selesai:
          type: string
          format: date
          description: Event end date (WIB local YYYY-MM-DD)
          example: '2026-05-12'
        jam_mulai:
          type: string
          description: WIB 24-hour HH:MM:SS (time serialized with seconds)
          example: '09:00:00'
        jam_selesai:
          type: string
          description: WIB 24-hour HH:MM:SS (time serialized with seconds)
          example: '21:00:00'
        lokasi:
          type: string
          description: Event venue/location name
          example: Alun-alun Purwokerto
        status:
          type: string
          enum: [draft, published, berlangsung, selesai]
          example: published
        stand_id:
          type: string
          nullable: true
          description: Assigned stand (e.g. 'A-3')
          example: A-3
        posisi_event:
          type: string
          nullable: true
          description: Position/role within the event
          example: Stand A-3
        status_request:
          type: string
          enum: [pending, pending_change, approved, rejected]
          example: approved
        assigned_by:
          type: string
          enum: [admin, self]
          example: admin
        created_at:
          type: string
          format: date-time
          example: '2026-05-01T08:00:00Z'
...
    Artisan:
      type: object
      required: [id, slug, username, nama_usaha, pemilik, email, no_hp, kota, kategori_usaha, status, tanggal_daftar, komisi_persen]
      properties:
        id:
          type: string
          example: art-0001
        slug:
          type: string
          description: URL-safe slug from nama_usaha
          example: batik-budi
        username:
          type: string
          pattern: '^[a-z0-9][a-z0-9-]{1,30}[a-z0-9]$'
          description: User-chosen handle for public profile URL (e.g. /profile/{username}).
          example: batik-budi
        nama_usaha:
          type: string
          description: Business/store name
          example: Batik Budi
        pemilik:
          type: string
          description: Owner/proprietor name
          example: Budi Santoso
        email:
          type: string
          format: email
          example: budi@example.com
        no_hp:
          type: string
          example: '081234567890'
        kota:
          type: string
          example: Purwokerto
        deskripsi:
          type: string
          example: Produsen batik lokal Banyumas
        foto_url:
          type: string
          nullable: true
          example: https://storage.example.com/foto/art-0001.jpg
        cover_url:
          type: string
          nullable: true
          example: null
        qris_url:
          type: string
          nullable: true
          description: QRIS payment image URL
          example: https://storage.example.com/qris/art-0001.png
        qris_updated_at:
          type: string
          format: date-time
          nullable: true
          description: |
            Set automatically by trg_artisans_qris_ts when qris_url changes.
            UTC ISO 8601. Read-only — backend should not accept in PATCH body.
          example: '2026-04-01T08:00:00.000Z'
        kategori_usaha:
          type: array
          items:
            type: string
          description: UMKM 9 options (NOT 'subsektor')
          example: ['F&B / Kuliner', 'Kriya']
        status:
          type: string
          enum: [aktif, pending, suspended, rejected]
          example: aktif
        tanggal_daftar:
          type: string
          format: date
          description: WIB local date YYYY-MM-DD
          example: '2026-01-15'
        komisi_persen:
          type: number
          description: Commission percentage set by admin
          example: 10
        total_penjualan:
          type: number
          description: Computed running total (rupiah)
          example: 5400000
        komisi_terkumpul:
          type: number
          description: Computed commission (rupiah)
          example: 540000
        internal_notes:
          type: string
          description: |
            **Admin-only** free-text metadata. NEVER returned by /api/public/profiles/{slug} or /api/artisan/me.
          example: Awaiting QRIS verification.
...
    Kas:
      type: object
      required: [id, jenis, kategori, metode, nominal, tgl]
      description: Read-only view of artisan cash flow (admin sees, cannot edit)
      properties:
        id:
          type: string
          example: kas-0001
        jenis:
          type: string
          enum: [masuk, keluar]
          description: Cash flow direction
          example: masuk
        kategori:
          type: string
          description: Transaction category
          example: Penjualan
        pelanggan:
          type: string
          nullable: true
          example: Ibu Sari
        barang:
          type: string
          nullable: true
          example: Kain Batik Cap
        qty:
          type: number
          example: 2
        metode:
          type: string
          enum: [tunai, qris]
          description: Payment method — strict whitelist per DB CHECK
          example: qris
        ket:
          type: string
          description: Notes/remarks
          example: Pesanan reguler
        nominal:
          type: number
          description: Amount in rupiah
          example: 300000
        tgl:
          type: string
          format: date
          description: WIB local date YYYY-MM-DD
          example: '2026-04-28'
        saldo_after:
          type: number
          description: Computed running balance (rupiah)
          example: 1500000
        bukti_url:
          type: string
          nullable: true
          example: null
        created_at:
          type: string
          format: date-time
          description: UTC ISO 8601, set on INSERT
          example: '2026-04-01T08:00:00.000Z'
        updated_at:
          type: string
          format: date-time
          description: UTC ISO 8601, auto-set by set_updated_at trigger on UPDATE
          example: '2026-04-15T10:30:00.000Z'
    Riwayat:
      type: object
      required: [id, pelanggan, barang, qty, total, metode, tgl]
      description: Read-only view of artisan sales history
      properties:
        id:
          type: string
          example: rwy-0001
        pelanggan:
          type: string
          description: Customer name
          example: Ibu Sari
        barang:
          type: string
          description: Product sold
          example: Kain Batik Cap
        qty:
          type: number
          example: 2
        total:
          type: number
          description: Total amount in rupiah
          example: 300000
        metode:
          type: string
          enum: [tunai, qris]
          description: Payment method — strict whitelist per DB CHECK
          example: qris
        tgl:
          type: string
          format: date
          description: WIB local date YYYY-MM-DD
          example: '2026-04-28'
...